+++ src/test/java/uk/gov/hmcts/juror/api/moj/service/CompleteServiceServiceImplTest.java
@@ -651,118 +651,7 @@ class CompleteServiceServiceImplTest {
         }
     }
 
-    @Nested
-    @DisplayName("public List<CompleteJurorResponse> search(JurorPoolSearch search)")
-    class Search {
-        @Test
-        @SuppressWarnings("PMD.NcssCount")
-        void positiveTypical() {
-            JurorDetailsDto jurorDetailsDto1 = mock(JurorDetailsDto.class);
-            when(jurorDetailsDto1.getJurorNumber()).thenReturn("111111111");
-            when(jurorDetailsDto1.getPoolNumber()).thenReturn("2222222222");
-            when(jurorDetailsDto1.getFirstName()).thenReturn("FNAME1");
-            when(jurorDetailsDto1.getLastName()).thenReturn("LNAME1");
-            when(jurorDetailsDto1.getPostCode()).thenReturn("POSTCODE1");
-            when(jurorDetailsDto1.getCompletionDate()).thenReturn(LocalDate.of(2023, 1, 1));
 
-            JurorDetailsDto jurorDetailsDto2 = mock(JurorDetailsDto.class);
-            when(jurorDetailsDto2.getJurorNumber()).thenReturn("111111112");
-            when(jurorDetailsDto2.getPoolNumber()).thenReturn("2222222223");
-            when(jurorDetailsDto2.getFirstName()).thenReturn("FNAME2");
-            when(jurorDetailsDto2.getLastName()).thenReturn("LNAME2");
-            when(jurorDetailsDto2.getPostCode()).thenReturn("POSTCODE2");
-            when(jurorDetailsDto2.getCompletionDate()).thenReturn(LocalDate.of(2023, 1, 2));
-
-            JurorDetailsDto jurorDetailsDto3 = mock(JurorDetailsDto.class);
-            when(jurorDetailsDto3.getJurorNumber()).thenReturn("111111113");
-            when(jurorDetailsDto3.getPoolNumber()).thenReturn("2222222224");
-            when(jurorDetailsDto3.getFirstName()).thenReturn("FNAME3");
-            when(jurorDetailsDto3.getLastName()).thenReturn("LNAME3");
-            when(jurorDetailsDto3.getPostCode()).thenReturn("POSTCODE3");
-            when(jurorDetailsDto3.getCompletionDate()).thenReturn(LocalDate.of(2023, 1, 3));
-
-            SecurityContextHolder.getContext().setAuthentication(
-                new BureauJwtAuthentication(List.of(),
-                    TestUtils.createJwt("415", "COURT_USER", "0"))
-            );
-
-            JurorPoolSearch poolSearch = JurorPoolSearch.builder()
-                .jurorNumber("1234")
-                .build();
-
-            PaginatedList<JurorDetailsDto> result = new PaginatedList<>();
-            result.setData(List.of(jurorDetailsDto1, jurorDetailsDto2, jurorDetailsDto3));
-            doReturn(result).when(jurorPoolService).search(poolSearch);
-
-            PaginatedList<JurorDetailsDto> responses = jurorPoolService.search(poolSearch);
-
-            assertThat(responses).isNotNull();
-            List<JurorDetailsDto> data = responses.getData();
-            assertThat(data).isNotNull().hasSize(3);
-            JurorDetailsDto response1 = data.get(0);
-            assertThat(response1).isNotNull();
-            assertThat(response1.getJurorNumber()).isEqualTo("111111111");
-            assertThat(response1.getPoolNumber()).isEqualTo("2222222222");
-            assertThat(response1.getFirstName()).isEqualTo("FNAME1");
-            assertThat(response1.getLastName()).isEqualTo("LNAME1");
-            assertThat(response1.getPostCode()).isEqualTo("POSTCODE1");
-            assertThat(response1.getCompletionDate()).isEqualTo(LocalDate.of(2023, 1, 1));
-
-            JurorDetailsDto response2 = data.get(1);
-            assertThat(response2).isNotNull();
-            assertThat(response2.getJurorNumber()).isEqualTo("111111112");
-            assertThat(response2.getPoolNumber()).isEqualTo("2222222223");
-            assertThat(response2.getFirstName()).isEqualTo("FNAME2");
-            assertThat(response2.getLastName()).isEqualTo("LNAME2");
-            assertThat(response2.getPostCode()).isEqualTo("POSTCODE2");
-            assertThat(response2.getCompletionDate()).isEqualTo(LocalDate.of(2023, 1, 2));
-
-            JurorDetailsDto response3 = data.get(2);
-            assertThat(response3).isNotNull();
-            assertThat(response3.getJurorNumber()).isEqualTo("111111113");
-            assertThat(response3.getPoolNumber()).isEqualTo("2222222224");
-            assertThat(response3.getFirstName()).isEqualTo("FNAME3");
-            assertThat(response3.getLastName()).isEqualTo("LNAME3");
-            assertThat(response3.getPostCode()).isEqualTo("POSTCODE3");
-            assertThat(response3.getCompletionDate()).isEqualTo(LocalDate.of(2023, 1, 3));
-
-            verify(jurorPoolService, times(1)).search(poolSearch);
-        }
-
-        @ParameterizedTest
-        @NullSource
-        @EmptySource
-        void negativePoolsNotFound(List<JurorDetailsDto> data) {
-            JurorPoolSearch poolSearch = JurorPoolSearch.builder()
-                .jurorNumber("123")
-                .jurorStatus(IJurorStatus.COMPLETED)
-                .build();
-
-            SecurityContextHolder.getContext().setAuthentication(
-                new BureauJwtAuthentication(List.of(),
-                    TestUtils.createJwt("415", "COURT_USER", "0"))
-            );
-            PaginatedList<JurorDetailsDto> response = new PaginatedList<>();
-            response.setData(data);
-            doReturn(response).when(jurorPoolService).search(poolSearch);
-
-
-            MojException.NotFound exception = assertThrows(MojException.NotFound.class,
-                () -> jurorPoolService.search(poolSearch),
-                "Exception should be thrown");
-
-            assertThat(exception).isNotNull();
-            assertThat(exception.getCause()).isNull();
-            assertThat(exception.getMessage()).isEqualTo(
-                "No complete juror pools found that meet your search criteria.");
-
-
-            verify(jurorPoolService, times(1)).search(poolSearch);
-
-        }
-
-
-    }
 
     @Nested
     @DisplayName("public void uncompleteJurorsService(String jurorNumber, String poolNumber)")
